Joe pid rotate and testing June 27th (#191)
* only rotates xyz, not rpy

xyz need to be rotated, this is odom->base_link

torques don't need to be rotated, that would create a new axis, not
transform between existing ones

* removes angular velocity from robot localization

this page
https://docs.ros.org/en/melodic/api/robot_localization/html/preparing_sensor_data.html

says "If the odometry provides both orientation and angular velocity,
fuse the orientation."

* changes acceptable dist to 0.2

before it was too small and would get stuck

* move around is now relative and has more points

* updates pre-qual mission

lowk this is untested but should work

we now spin and can pre-qual at depth

* rotates the cmd_wrench before sending it

why? becuase the cmd_wrench is defined with respect to odom, but the thruster's expect a msg with
respect to base_link! If the sub is rotated these two frames mis-align and the sub acts in ways
that seem random or incorrect

* new pid gains for z (also corrects sign errors)

---------

Co-authored-by: Joseph Handsome <jgoodman1@ufl.edu>
Co-authored-by: Joseph Handsome <nottellingu@gmail.com>